Moved the producer resolver to its own file
