+++ server/schemas/queries/index.js
@@ -1,3 +1,4 @@
 export { getAllUsers, getSpecificUser, getMe } from "./user.js";
 export { getVarieties } from "./variety.js";
 export { getRegions } from "./region.js";
+export { getProducers } from "./producer.js";
