+++ server/schemas/resolvers.js
@@ -10,7 +10,8 @@ import {
     getSpecificUser,
     getMe,
     getVarieties, 
-    getRegions 
+    getRegions,
+    getProducers
 } from "./queries/index.js";
 
 import { signToken } from "../utils/auth.js";
@@ -27,23 +28,7 @@ const resolvers = {
 
         regions: getRegions,
 
-        producers: async (parent, { id, name }, context) => {
-            if (context.user) {
-                let producers = [];
-                if (id && name) {
-                    throw new GraphQLError("Please provide either an ID or a Name when searching producers", { extensions: { code: "BAD_REQUEST" } });
-                } else if (id) {
-                    producers.push(await Producer.findById(id));
-                } else if (name) {
-                    producers.push(await Producer.find({ name }));
-                } else {
-                    producers = Producer.find({});
-                }
-
-                return producers;
-            }
-            throw new GraphQLError("Please login", { extensions: { code: "UNAUTHORISED" } });
-        },
+        producers: getProducers,
 
 
         // Wine search - Type determines any fixed category
